tmo-restoran-panel

+++ src/main/java/hamit/demir/model/dto/StokKalemi/StokKalemiResponse.java
@@ -2,12 +2,13 @@ package hamit.demir.model.dto.StokKalemi;
 
 import hamit.demir.model.entity.Birim;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record StokKalemiResponse(
         Long id,
         String ad,
-        Integer miktar,
+        BigDecimal miktar,
         Birim birim,
         Boolean aktif,
         LocalDateTime olusturmaTarih,
