+++ src/main/java/hamit/demir/model/dto/StokKalemi/StokKalemiSaveRequest.java
@@ -4,11 +4,14 @@ import hamit.demir.model.entity.Birim;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record StokKalemiSaveRequest (
         @NotBlank String ad,
-        @NotNull Integer miktar,
+        @NotNull BigDecimal miktar,
         @NotNull Birim birim,
         @NotNull Boolean aktif
+
 
 ){
 }
