+++ src/main/java/hamit/demir/model/dto/StokKalemi/StokKalemiUpdateRequest.java
@@ -4,11 +4,13 @@ import hamit.demir.model.entity.Birim;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record StokKalemiUpdateRequest(
 
         @NotNull Long id,
         @NotBlank String ad,
-        @NotNull Integer miktar,
+        @NotNull BigDecimal miktar,
         @NotNull Birim birim,
         @NotNull Boolean aktif
 ) {
